[otap-dataflow] Micro optimizations for the view implementations (#689)
## Changes
A few micro-optimizations for the view implementations
- Use `&str` instead of `Cow<'a,str>` 
- Add `#[inline]` to all the methods that are used when encoding data to
OTAP
- For the bytes implementation, update the `advance_to_find_field`
method to
  - Enter the loop only if the field is not already cached
- Use a simpler check `field == field_num` instead of calling
`get_field_range` to know when to return

I made the first two changes to both the prost and the bytes
implementation to have a fair comparison.

For the bytes representation, these changes showed a ~5% improvement in
the traversal benchmarks and ~3.5% for the encoding benchmarks.