[query-engine] Introduce PipelineExpressionBuilder (#687)
Related to #684

## Changes

* Introduce `PipelineExpressionBuilder` and refactor KQL parser to use
it

## Details

The main goal here is to introduce `PipelineExpression.optimize()` which
is invoked as the last step before the `PipelineExpression` is returned.
This gives us a spot to introduce constant folding and other
optimizations, validations, etc., when queries are being parsed or when
pipelines are manually constructed (if someone wanted to do that).