[query-engine] Re-export Parser API in parser crates (#707)
## Changes

* Re-export `Parser` & `ParserError` in kql and ottl parser crates

## Details

We have this
[Parser](https://github.com/open-telemetry/otel-arrow/blob/4f55ba7eb4abfe5f57177d15603df0e86b23f47a/rust/experimental/query_engine/parser-abstractions/src/parser.rs#L7)
trait which is the entry point for parsing KQL or OTTL queries.

In order to call these traits users have to bring them into scope which
currently means depending on the `parser-abstractions` crate. What this
change does is makes them part of the public API in the KQL & OTTL
crates so users don't need to worry about `parser-abstractions`.

I feel like this is a good thing to do but I'm also new to Rust 😄 It may
be something which gets reverted later but I thought it was worth trying
out.