[query-engine] Add layout for new recordset engine and primitives (#747)
## Changes

* Adds a new project engine-recordset2
* Adds primitives used by the engine to execute expressions

## Details

The current recordset engine is really a proof-of-concept. It contains
its own expression tree and everything is tied deeply to the
implementation. This new engine works more or less in the same way
except it run expressions from the dedicated project. The goal is to
build this up and then remove the poc version.