feat: add generic wrappers support OTLP view generation (#753)
Generating reference wrappers and iterators for OTLP views is tedious,
so this change introduces generic wrappers and iterators to cut the
tedium (this makes life much easier in the next PR I'm going to submit
which adds OTLP protobuf struct to OTAP conversion for metrics). I've
updated the recently added views for traces (but not logs or common) to
make use of these wrappers.

Co-authored-by: Laurent Quérel <l.querel@f5.com>