[WIP] feat: add OTLP protobuf to OTAP conversion for metrics
This change is similar to previous work done for logs (#625) and traces (#691) but metrics are larger and messier.

There are a few bits that I still need to clean up, but given the size of the change, I wanted to get it front of folks for reviewing sooner rather than later. Specifically, I'd like to:
* add some tests
* double check OTAP storage semantics in a few places where I found them confusing
* use the recently added `append_many` functionality to replace some `(0..metric_count).for_each` calls.