First version of the engine capable of creating and executing a pipeline from a configuration (#532)
This PR represents the first step where the dataflow engine is able to
run an end-to-end pipeline from a configuration. In its current version,
the engine still has several limitations (which will be addressed in
later PRs):
- No controller automatically starting as many engines as there are
cores.
- No support for hyper-edges with a broadcast strategy.
- No support for output ports.

A pipeline configuration is passed to the `PipelineFactory::build`
method, which then performs the following sequence:
- Create all runtime nodes based on the pipeline configuration.
- Analyze both the local vs shared nature of each pair of connected
nodes and the nature of the hyper-edges between them to determine the
best channel type.
- Assign channels to the source nodes and their destination nodes based
on the previous analysis.
- Return the resulting DAG as a `RuntimePipeline`.

This `RuntimePipeline` can then be started with a single-threaded tokio
runtime.

This PR demonstrates that our dataflow engine can be:
- configured from a configuration file
- composed of nodes (receivers, processors, exporters) that are a mix of
local and shared nodes
- able to create a DAG connecting the nodes with different types of
channels depending on the DAG topology, the nature of the nodes, and the
configuration of the hyper-edges connecting them
- started with a single-threaded tokio runtime and can process messages
through the pipeline

Note: I think this PR is ready for review even though I still need to
resolve a few build issues. In particular, there is a Rust formatting
problem with generated code in otel-arrow-rust.