[PerfTest] Backend performance (#775)
Fixes #773.

This PR converts the perf test grpc backend service over to use
grpc.aio. This alleviates an observed bottleneck with high thread counts
on the sending side and "sending queue is full" error messages in the
otel collector.

Before / After performance with same configuration settings:

| metric_name                       |   threadpool |      asyncio |
|:----------------------------------|-------------:|-------------:|
| Total logs attempted              |  5.9663e+07  |  7.34805e+07 |
| Logs successfully sent by loadgen |  5.9663e+07  |  7.34805e+07 |
| Logs failed at loadgen            |  0           |  0           |
| Logs received by backend          |  5.76995e+07 |  7.34805e+07 |
| Logs lost in transit              |  1.9635e+06  |  0           |
| Logs receive rate (avg)           |  1.40858e+06 |  2.08576e+06 |
| Total logs lost                   |  1.9635e+06  |  0           |
| Percentage of logs lost           |  0.0329098   |  0           |

I'm unable to push data fast enough to reproduce the condition with
asyncio on my laptop so this is likely good enough for now.

Added some light testing as well.

---------

Co-authored-by: Laurent Quérel <l.querel@f5.com>