Document load balancing challenges and solutions (#736)
I created this document as a follow-up to a [Slack conversation I had
with
Jake](https://cloud-native.slack.com/archives/C07S4Q67LTF/p1752601543332769).
The goal is to explain the challenges related to OTAP load balancing and
to explore possible solutions (some of which have already been explored,
particularly on the client side).