Fix DebugExporter and detection of blocking function in async context (#785)
The `DebugExporter` was using blocking functions that were blocking the
async runtime when several tasks were running in parallel (e.g.
`std::io::Write`). I replaced the blocking functions with their Tokio
equivalents.

In addition, an `info` async method has been added to the effect
handlers to emit messages to stdout in a non-blocking manner.

I also added a script to detect this kind of issue in the rest of the
code, and there are many occurrences of this problem!
I encourage you to look at the output of this script below and fix the
corresponding issues. I will add this script to our `cargo xtask check`
once we no longer have any problems, so that it can be detected in the
build in the future.

<img width="1170" height="690" alt="image"
src="https://github.com/user-attachments/assets/9e3cb332-a58c-4f80-ad28-d138c87bc3a4"
/>