fix: flakey shutdown timeout test (#801)
fixes: https://github.com/open-telemetry/otel-arrow/issues/759

Fixes the flakey shutdown timeout test in parquet_exporter.

Before this change, the test was passing a
`ControlMsg::Shutdown(Duration::ZERO)`, which would cause the channel to
send a shutdown message to the exporter without trying to drain pdata.
Most of the time, this would mean that calling `writer.flush()` would
instantly finish because there were no buffered writes, and we wouldn't
hit the timeout error. But occasionally, the timeout future would
complete first for some reason and the test would fail.

Obviously we want to make this more reliable, while also testing the
proper timeout behaviour. To do this, we need to invoke the drain
mechanism to buffer some writes, in order that parquet exporter's
`writer.flush()` method takes longer than the shutdown timeout.

To do this, we need a reliable way to invoke the drain mechanism. We do
this by passing `ControlMsg::Shutdown(Duration::from_nanos(1))`, which
will cause the `MessageChannel` to initiate drain behaviour. Also, we
augment the behaviour at start of the `MessageChannel` control loop
where if `shutting_down_deadline` is `Some`, we always initiate draining
first. This gives a more reliable shutdown experience, because otherwise
there's unknown behaviour related to how long the control loop iteration
takes, how long the shutdown timeout is, and whether or not we attempt
to drain pdata before sending the shutdown control message.

Also, in order to validate that the exporter returns the expected result
the test harness now passes the returned `Result` into the validation
procedure callback.