feat: OTLP receiver that converts proto bytes to OTAP (#726)
Adds an `otlp_receiver` implementation (in the `otap` module) that can
receive the OTLP GRPC messages, but converts the proto bytes into OTAP.
The idea is that this would be the entry-point to an OTAP telemetry
pipeline that is receiving OTLP telemetry from GRPC.

By implementing Tonic's `Codec`
https://docs.rs/tonic/latest/tonic/codec/index.html traits, we can get
access to actual bytes of the GRPC message to do the decoding.

Currently only logs are supported, but it should be very straight
forward to add capability handle metrics & traces once we have the proto
bytes view implementations for these signal types & the OTAP encoding
implementation.

Related: https://github.com/open-telemetry/otel-arrow/issues/721