[Perf Test] support linux psutil handling, docker dependency (#807)
This fixes an issue with the different behaviors of psutil.cpu_times()
between linux and mac/windows, and removes dependency on docker for
configurations which don't use it.

- Manually itterate over child processes and aggregate system/user cpu
times for consistent output in mac/linux.
- Defer docker client initialization so that auto-registered hooks don't
raise exceptions with no docker compatible api is found.
- Revert to previous backend exit signal handling which is hanging
indefinitely in linux until it can be triaged.

---------

Co-authored-by: Laurent Quérel <l.querel@f5.com>