Use new `OtapPata` type in `ParquetExporter` (#810)
Fixes https://github.com/open-telemetry/otel-arrow/issues/809

Use the new `OtapPData` type in `ParquetExporter`. Before this it was
generic over anything that could produce `OtapArrowRecords` & `batch_id:
i64`. This change enables us to use create the `ExporterFactory` for
this component.

Also, it was unnecessary to have the `batch_id` in the pdata. For
`ParquetExporter`, this was just being used as a high watermark when
figuring out which record batches the writer should flush, so it's
sufficient to just use a locally incrementing counter (since the writer
is also local to the `start` method).

Also adds some test for deserializing config. These kind of serve as an
example for how the JSON config should look (it's sometimes not always
clear, especially when it comes to enum variants).

---------

Co-authored-by: Laurent Quérel <l.querel@f5.com>