feat: add otlp grpc exporter for otap pipelines (#819)
Resolves https://github.com/open-telemetry/otel-arrow/issues/815

Adds an implementation of the OTLP GRPC exporter that can be used in
OTAP pipelines. It expects to receive the request objects already
serialized (as it exists in the `OtlpProtoBytes` variant inside the
`OtapPdata` type).

This change moves the otlp grpc server code in the `otap::grpc::otlp`
module into a new module called `server`, and adds a new module called
`client` that has the grpc client implementation along with the thin
codec implementation that does no serialization of the request. The
client implementation is generic of the logs/metrics/traces service.

I wanted to reuse the code in `otap_df_otlp::mock` in my tests, so this
code also adds a new feature to that crate to expose those.

---------

Co-authored-by: Laurent Quérel <l.querel@f5.com>