[query-engine] Static value refactor (#854)
## Changes

* Rename `Value` and `ValueMut` -> `StaticValue` & `StaticValueMut`
* Re-introduce `Value` without `'static` requirement
* Use `From` \ `Into` for conversions instead of custom logic on
`ValueSource<T>`
* Use blanket implementations of traits where possible to allow removal
of a lot of code repating boilerplate things like `impl AsValue for
[SomeType] { ... }`

## Details

The main thing this addresses is the `'static` on `Value`...

```rust
enum Value<'a>
{
   Array(&'a (dyn ArrayValue + 'static))
   ...
}
```

The `'static` there means the thing implementing the trait must be
statically-known. It doesn't mean it needs a 'static` lifetime. One of
the restrictions for something to be known statically is that it can't
itself have a lifetime or contain any pointers.

What I'm working on is slice support in the engine. We want a scalar
expression which will take a slice of a string or an array. To do this
with decent perf you want the slice to hold a reference to the thing it
is slicing. Problem is, the `'static` prohibits this.