Fix pipeline perf test to run on bare metal runner (#698)
~Need to temporarily enable running the perf tests on PR that has the
label "performance" added to it. This is to make it easy for me to
investigate and fix the issues in the runner (We have permissions now to
use the new dedicated Oracle bare metal VMs, but the CI is failing for
other reasons that needs to be investigated - this makes the
investigation easier)~

Updates the action to install python and run the test on bare metal. The
alternative is to use docker, but I think this would likely give maximum
performance.
(Will check with machine admins separately to see if they can install
python on the runners, so we don't need to do it in gh action)

Example run on baremetal:

https://github.com/open-telemetry/otel-arrow/actions/runs/16819130527/job/47642481063

---------

Co-authored-by: Joshua MacDonald <jmacd@users.noreply.github.com>