fix: correctness fixes for OTLP to OTAP conversion for traces (#865)
Working through the metrics version of this code taught me a lot which
gave me the opportunity to fix a number of issues in the traces version:
* we now store a `schema_url` for scopes separate from the one defined
for resources
* `Status` is now treated as a separate `StructArray` column rather than
an inlined pair of columns
* we now use the correct `Duration` datatype for the
`duration_time_unix_nano` field
* we no longer use dictionary encoding for `span_id`, `trace_id`, and
`parent_span_id`

In addition, I've centralized definitions for `TraceId` and `SpanId` so
they're defined once and only once.

Finally, I've removed `Default` impls for record batch builders. These
only existed because of a clippy lint (which I've disabled in
`encode::record`) and I think that they're a bit harmful. The builders
are heavy structs which make a lot of allocations on construction; I
think that's a somewhat unexpected property for a type that implements
`Default` and I worry about folks using `impl Default` without realizing
the cost.