fix: OTAP decoding handles plain encoded log IDs (#898)
part of: #878 

When we originally wrote the OTAP -> OTLP decoding code, we made the
assumption that the ID column was always delta encoded (because this is
what the golang exporter produces). This assumption no longer holds, as
we use this code to convert between OTAP -> OTLP in OtapPdata, but our
OTAP encoder produces ID columns that are not delta encoded (e.g. plain
encoded).

We'll need to go change the OTLP decoder code to handle this in many
places. For now, this PR just handles this for logs in order to unblock
the work on syslog receiver (see discussion
[here](https://github.com/open-telemetry/otel-arrow/pull/861/files#r2255190720)).
Will fix for metrics and traces in followup PRs.

Also fixes: https://github.com/open-telemetry/otel-arrow/issues/481

---------

Co-authored-by: Laurent Quérel <l.querel@f5.com>