fix: OTAP decoding handle plain encoded attribute IDs (#913)
Part of: #878 

We have some bugs in the OTAP -> OTLP decoding code where the OTLP
decoding expects all the IDs & Parent IDs to be encoded as they are
produced by the Golang Collector. For attribute Parent IDs, this means
that we expect the Parent IDs to use the transport-optimized quasi-delta
encoding. However, our OTAP encoder (which is used when converting OTAP
Pdata representation), always produces plain encoded IDs & parent IDs.

This PR adds support to the OTLP decoder to handle these plain encoded
parent IDs. To do this, the attribute store will check the metadata on
the parent ID column and not use the parent ID decoder if the ID is
plain encoded.

This PR also adds a helper method called `with_plain_encoding()` to the
Arrow `Field` type via an extension trait, which adds the correct field
metadata. This cuts down the verbosity of having to set the field
encoding using `HashMap::from_iter` in many places (especially in test
code).

---------

Co-authored-by: Utkarsh Umesan Pillai <66651184+utpilla@users.noreply.github.com>