replace `rename_attributes` function with `transform_attributes` w/ delete (#938)
Part of #813 

Replaces the `rename_attributes` function with a new
`transform_attributes` function that can replace attribute keys and
delete attributes by key in a single operation. The motivation by doing
these in bulk is to minimize the number of times we may need to
materialize immutable arrow buffers when doing multiple attribute
transforms.

The one outstanding TODO on this is to remove optional columns that,
because some rows were deleted, now contain entirely null or default
values. I can probably do this in a followup

---------

Co-authored-by: Laurent Quérel <l.querel@f5.com>