Remove leftover otel-arrow-rust workflow and add license header checker (#954)
Noticed that Renovate was trying to keep workflow files up-to-date
inside `otel-arrow-rust` (#948) which aren't actually executed anymore
and a leftover from when that code was in its own repository.

This PR:
- Removes the old workflow file
- Bridges gap between old file and current workflows by adding
`license-header-check` job
- See https://github.com/korandoru/hawkeye?tab=readme-ov-file#usage for
more information

Note: This currently fails intentionally, want to confirm direction with
maintainers/contributors before actually applying.

The command `hawkeye format` can be used to automatically add the header
on every flagged file - maybe should be added to the `Makefile` as a
command shortcut?