Add transport optimized sorting & encoding to OTAP Pdata when converting to IPC Format (#972)
Fixes: #931 

Adds a new function to `OtapArrowRecords` called
`encode_transport_optimized` which converts the contained Arrow Record
Batches to be optimized for transport. This is done automatically by the
`Producer` when converting between `OtapArrowRecords` and
`OtapArrowBytes` (e.g. Arrow Record Batches -> IPC Serialized batches).

The "transport optimized" conversion involves sorting the record
batches, and then delta encoding the IDs. Because we can't have
sequences of negative deltas (which may be produced after sorting), the
ID columns are rewritten and then the parent ID column of child record
batches are mapped back to the new IDs.

The majority of the implementation is in a new module in
`otel-arrow-rust` located at `otap::transform::transport_optimized`.

---------

Co-authored-by: Utkarsh Umesan Pillai <66651184+utpilla@users.noreply.github.com>
Co-authored-by: Laurent Quérel <l.querel@f5.com>