Restore parquet exporter shutdown test and fix flakey implementation (#991)
Fixes: #967 

I think the source of the flakiness here was we'd enter into this:

https://github.com/open-telemetry/otel-arrow/blob/b4bb1d87434528b7ffedc6aa5c76b512f6358182/rust/otap-dataflow/crates/engine/src/message.rs#L221-L241

And depending on how long it took between when the test setup function
run, and when we invoked the validation function, we'd either hit the
deadline or the pdata first. If we hit the deadline first, no writes are
queued and the parquet exporter shuts down successfuly. This is not what
we're trying to test here.

To get around this, the test is now rewritten to manually drive the
control messages, and have the parquet exporter receiving them at the
same time. This way, we can have a better guarantee that there will be
unflushed writes queued when it receives the shutdown signal.