Upgrade to arrow 56.1 to improve adaptive schema handling in parquet exporter (#1006)
part of #863 

Improvements were made in the latest version of arrow-rs
(https://github.com/apache/arrow-rs/pull/8095 and
https://github.com/apache/arrow-rs/pull/8005) to improve parquet
writer's ability to handle logically equivalent data types for the same
column.

We need this b/c different OTAP batches might switch between Dictionary
encoding (with various key sizes) and native encoding for the same
record batch. The latest version of parquet is able to accept these
different array types for the same column.

A test is added to ensure the behaviour works as expected.

---------

Co-authored-by: Laurent Quérel <l.querel@f5.com>