Split Error and TypedError<T>; generalize NodeControlMsg<PData> (#1017)
Part of #509 

While working to implement the rest of the retry_processor support,
discovered a need to add `<PData>` to the `NodeControlMsg`. The `Nack`
message now includes an optional `pdata` field.

At the same time, most uses of Error did not require a `<T>`, so this
separates them and handles conversion to and from the `TypedError<T>`
form. The type is erased and the error message is stored as a string.

No functional changes except w.r.t. error structure and a new
`Option<Box<PData>>` in the Nack. While these changes are unrelated,
they have approximately the same impact and combining them was
relatively easy.

---------

Co-authored-by: Laurent Quérel <l.querel@f5.com>