Use Arrow row format when sorting for batching if possible (#1043)
Closes: #1026 

Improve batching performance by using the arrow row format for
multicolumn sort where possible.

We do this by creating a reusable function called
`otap::transform::sort_to_indices` and using this both in the batching
code and in the transport optimization encoding code (which is where
this was refactored from).