Avoid flattening dictionaries when unifying record batches during batching (#1042)
Closes: #1025 

When concatenating OTAP batches, there are two issues that are a bit
difficult WRT to dictionary columns:
- the combined record batch might overflow the dictionary due to more
keys having more cardinality than the key type can support
- received record batches may have different data types for the same
dictionary column (it depends on the size of the batch, and the
cardinality of the column when it was constructed).

Before this, we were using a simler logic of flattening all
dictionaries, however in doing this we lose the benefits of dictionary
compression.

With this PR, we no longer flatten the dictionaries. Instead, we attempt
to determine the correct key size to use and the cast the column for all
record batches to be the correct data type. We try to make the decision
about which key type to use in the most efficient way by considering the
total batch size, the possible cardinality of the columns and the
minimum dictionary key size.

---------

Co-authored-by: Laurent Quérel <l.querel@f5.com>