Improve pipeline draining (#1058)
In this PR, we no longer send a shutdown control message to all nodes in
a pipeline but only to the receivers. This forces the pipelines to stop
in the receiver-to-exporter direction. It's not the end of the story for
achieving graceful draining of the pipelines, but it's a good start.