Cleanup of the OTLP crate (#1069)
Most of the code in the `otlp` crate was no longer used, as it had been
migrated to the OTAP pipeline.
The code still in use (shared between OTLP and OTAP) was moved to the
OTAP crate.
Rationalization of `Cargo.toml` to use only one version of `tonic` and
`prost`.
Updated `README.md`.

Future PR: The next step is to migrate to `tonic/prost` 0.14.