fix issue decoding attribute key-values with special characters (#1077)
Fixes: #1060 
Fixes: #1070 

When we decoded an attribute, we weren't advancing the `RawKeyValue`'s
position in the buffer to the start of the next field, we'd just advance
it to the byte after the field's tag (e.g. where the len starts). Then
when we go to read the values, sometimes we'd parse each byte in the key
as the start of a new proto field as if it was the start of a new field
by treating it like a variant and advancing. For some key values, we'd
just happen to stumble across the start of the value and everything
would be fine. But depending on the bytes, sometimes we'd miss it and
decode the attribute as empty.

---------

Co-authored-by: Laurent Quérel <l.querel@f5.com>