Add configurable arrow IPC compression to otap exporter & align default compression/IPC stream options (#1085)
Closes: #1084 

Adds the ability to set the Arrow IPC compression in the OTAP exporter,
similar to what is configurable in configurable in the golang exporter:
```yaml
exporter:
  config:
    arrow:
      payload_compression: zstd / none
```
`ztd` is the default, same as golang collector:

https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/adbd65ed00ff447a96ce9bdce1320424e5bb420d/exporter/otelarrowexporter/internal/arrow/exporter.go#L62

This PR also changes the default gRPC compression to zstd, which is also
what is configured by default in the golang collector:

https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/adbd65ed00ff447a96ce9bdce1320424e5bb420d/exporter/otelarrowexporter/factory.go#L68-L69

So by default, we have the "double compression".

This PR also turns on dictionary delta encoding in the IPC stream
reader. This also aligns with what the go collector configures:

https://github.com/open-telemetry/otel-arrow/blob/18996ac489c7ed98f315b803275a3dc112844fc5/go/pkg/otel/arrow_record/producer.go#L403