feat: improve performance of dictionary unification code in batching (#1105)
We employ a few different tricks to improve performance here, mostly by
removing expensive operations from our inner loop:

* Originally, we were performing type dispatch for each element. Now, we
dispatch once per array.

* We used to check for null elements for every single index in the
array; now we employ `NullBuffer::valid_slices()` to get a contiguous
slice of valid elements; in the worst case (an alternating sequence of
null and not-null values), this devolves into the same performance we
had before, but in all other cases, this allows us to amortize
null-handling over a much larger set of values.

* For arrays that don't have any nulls, either because their data-type
is non-null or because there just happen to be no nulls present, we skip
null-handling altogether and operate on slices directly.

* We size `HashSet` capacity based on the total number of elements so
for small data sets, we should reduce allocation sizes.

* For small element types (those that fit within a `u32`), we skip the
`HashSet` of byte slices and opt for a `RoaringBitmap`.
[Roaring](https://arxiv.org/pdf/1603.06549) is a crate widely used in
database implementations to compactly describe sets of integers by
dynamically selecting a representation at runtime that includes dense
bit vectors, sparse sequences of integers, and intervals. The result is
fast and memory efficient. For anything `u32` sized or smaller, we just
convert the bits into a `u32`, stash it in a `RoaringBitmap` and then
rely on the bitmap's `len` method. The only caveat here is that the
flexible representation means that `RoaringBitmap::len()` does some real
work as opposed to `AHashSet::len()` which just reads a single value in
memory, so instead of checking for early termination on every array
value, we amortize it over every 256 array values.

This is not strictly a performance improvement, but I've also changed
`total_batch_size` to refer to the number of non-null items in a
sequence of columns. I think the math works but I'd appreciate extra
review around that change.

Note that the unified diff for this PR looks really confusing, so you
might benefit from switching to the side by side diff view.