Encode OTAP -> OTLP proto bytes directly (#1109)
part of #1095 
closes https://github.com/open-telemetry/otel-arrow/issues/1123

This PR adds an optimized mechanism for encoding OTAP -> OTLP proto
bytes directly. This implements the encoding only for Logs. Traces &
Metrics can happen in followup PRs. This PR also integrates this
optimized mechanism into the OTLP exporter.

In order to do the encoding in a single pass, we allocate a fixed sized
length encoding for some fields and zero-pad the varint. This is not
optimal from a storage perspective, and we could probably try to
optimize the size we guess in the future by inspecting the arrow data.

This PR also doesn't implement direct encoding for Map/Slice type
attributes, which are CBOR encoded. This can be done in a followup PR.
For now, we decode the CBOR to the prost struct and encode that.

Some quick performance testing was performed on a pipeline like
following:
- loadgen = 500,000 logs/sec
- batch size 8000
- single CPU core per pipeline
- machine = M4 macbook pro (my laptop)
```
                                                                
      ┌──────────────────────────────────┐                      
      │ load gen      core 1             │                      
      │                                  │                      
      │ ┌────────────┐     ┌───────────┐ │                      
      │ │ fake data  ┼────►│OTLP       ┼─┼───────┐              
      │ │ generator  │     │exporter   │ │       │              
      │ └────────────┘     └───────────┘ │       │              
      └──────────────────────────────────┘       │              
                                                 │              
                                                 │              
  ┌──────────────────────────────────────────────┘              
  │                                                             
  │  ┌──────────────────────────────────────────────┐           
  │  │    system under test          core 2         │           
  │  │                                              │           
  │  │┌────────┐    ┌────────────┐     ┌───────────┐│           
  │  ││OTLP    │    │  attribute │     │  OTLP     ││           
  └─►►│Receiver┼───►│  processor ┼─────►  Exporter ┼┼─────► ... 
     │└────────┘    └────────────┘     └───────────┘│           
     │                                              │           
     └──────────────────────────────────────────────┘           
                                                                
```
Results - CPU average utilization:
- `main` 65.1%
- `albert/1095`44.3%
- ~20% decrease

---------

Co-authored-by: Utkarsh Umesan Pillai <66651184+utpilla@users.noreply.github.com>
Co-authored-by: Laurent Quérel <l.querel@f5.com>