[perf] Avoid unnecessary vec allocation for merged transforms (#1180)
## Changes
- In the current implementation, we always allocate a `vec` when merging
transform ranges (even if there is no merge required for example if
there are either only replacement ranges or only delete ranges)
- With this PR, we append the `KeyTransformRangeType` data from the
get-go and store them in `KeyReplacementPlan` and `KeyDeletePlan`. This
allows us to simply refer to these fully self-contained ranges in the
`merge_transform_ranges` method if there is no merge required.