[query-engine] Improve validation in KQL accessor expressions when source is referenced directly (#1210)
## Changes

* Improve the schema validation in KQL accessor expressions when
`source` is referenced directly

## Details

We have schema validation in the KQL parser. When schema is defined
we'll use it to resolve types and detect invalid usage when possible.

Currently we only do schema validation when `source` is referenced
implicitly. For example `extend c = SomeField`. What this PR does is
extend that validation to also run when `source` is referenced
explicitly for example `extend c = source.SomeField`.